1. Unity FirstPersonCotroller Assets
Setting Up The Input System

The new first person controller asset uses the new Unity input system, which we do not currently cover during this section of the course.

However, you can still utilize the older input system for all of the actions that we create for our Zombie Runner game, such as switching and shooting our weapons.

To ensure that Unity can read both the new and old input types, you need to:

Go to Edit > Project Settings

Click on Player in the left-hand menu

Scroll down to Other Settings > Configuration

Check that the Active Input Handling* dropdown is set to Both.