3. Using NavMeshAgent for AI
Creating a Enemy placeholder where I gave it a component called Nav Mesh Agent.

Nav Mesh Agent lets the object navigate the scene without hitting static objects.

Set the settings of the NavMesh for object traversal.

Created EnemyAI script to handle target of the player.

Used using.Unity.AI namespace to access NavMeshAgent.

On Start(), declared:

navMeshAgent = GetComponent<NavMeshAgent>();

On Update(),

navMeshAgent.SetDestinatio();

set it to the position of the target(Player). so that it follows the player.

By using the NavMeshAgent, it will traverse and follow the player without interacting with any static objects.